backend/requirements.txt
Flask
flask-cors
whisper
pydub
torch
python-dotenv
uvicorn
gunicorn